Adding support for Marked 2.4's window and id handler in URL's
I've changed this macro from using a shell script and a temporary UUID to use the new `window` and `id` handler in Marked 2.4's URL scheme.

The macro remembers frontmost app and window and saves them to variables, which are then used as window and id in Marked. (`marked://preview?id=AppName&window=WindowName&text=PreviewText`)

This way you should be able to execute this macro in the same window again and have it *update* the same window in Marked, *without* opening a new window every time.