New feature: suppress <p> tag on conversion.
This is a nice feature for bloggers. Working with a system like WordPress often means that we have to write in HTML, but outputting everything with `p` tags around it makes the whole text somewhat cumbersome to read. Therefore you can now set to *not* output these tags.

This means writing this:

    This *is* **a** [test](http://zcasting3000.com).

Will output:

    This <em>is</em> <strong>a</strong> <a href="http://zcasting3000.com">test</a>.

Instead of:

    <p>This <em>is</em> <strong>a</strong> <a href="http://zcasting3000.com">test</a>.</p>

Note that this new feature is **OFF** by default because Multimarkdown does it that way, and I wouldn't want to mess with the mmd default. To set this open the `Helper - Settings` macro and change the variable `MMD__Suppress p Tag` from `NO` to `YES` and click `Try`, to set the variable.